discussions/serializers.py
from rest_framework import serializers
from .models import (
    Discussion,
    Like,
)


# serializers.py

class DiscussionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Discussion
        fields = ['id', 'user', 'text', 'image', 'hashtags', 'created_on']


class LikeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Like
        fields = ['id', 'user', 'discussion', 'created_on']
